[105702]Create and initialize the HTML hidden field for 'client-side time' on the client-side.
This field wasn't added to the login form in omniauth-ldap repository in the previous
fix (297020aed920e35ef68).

Instead of duplicating this field in all the current OmniAuth
authentication forms scattered in the different omniauth-* identity
strategy git repositories, our momentary decision is to keep this
client-side field in the client-side js.